Split the occupancy realize pass by topic
Realize.cs had grown to 2821 lines and 138 members, covering everything from
resolving a page's properties to seeding a newly created citizen's age. Nothing
about that is one subject, and at that size the file is only navigable by
search.

Split along the seams the code already had, into eleven partial files of 190 to
430 lines each - resolve, staging, property, bootstrap, household, vehicles,
citizens, create, move-in, support - each carrying a note saying what belongs in
it. Realize.cs keeps the shared state and the entry point, and points at the
rest.

Pure code movement: the members were compared byte for byte before and after,
and come out identical and in the same order.
